requirements.txt
selenium==4.4.3
django==4.1
djangorestframework-3.13.1
django-taggit==3.0.0
django-material-admin-1.8.6